requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
Click==7.0
Django==2.2.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2-binary==2.8.2
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
wrapt==1.11.2
